新增订单和产品功能

+++ src/main/java/com/makenv/service/OrderService.java
@@ -6,9 +6,9 @@ import java.util.Map;
 
 public interface OrderService {
 
-    ServerResponse pay(long orderNo, int userId, String path);
+    ServerResponse pay(Long orderNo, Integer userId, String path);
     ServerResponse aliCallback(Map<String, String> params);
-    ServerResponse queryOrderPayStatus(int userId, long orderNo);
-    ServerResponse createOrder(int userId, int shippingId);
-    ServerResponse cancel(int userId, int orderNO);
+    ServerResponse queryOrderPayStatus(Integer userId, Long orderNo);
+    ServerResponse createOrder(Integer userId, Integer shippingId);
+    ServerResponse cancel(Integer userId, Long orderNO);
 }
